+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added 
 - gameVersion, rep, sid & label to logs
 - DamageStateLog
+- level to battler log
 - notification when your rep is high enough for an upgrade
 
 ### Removed